selection: Move whole-file batch discard
The single-file discard-to-batch support also carried selected binary and
text-deletion batch helpers.

Those helpers act on whole selected file changes, while file-scope support
should stay focused on path-scoped fallback orchestration.

This commit moves the whole-file batch discard helpers into selection
support and leaves file-scope code to import them.

A following tests commit will pin the new boundary.
